u.2u4u.com.cn can't use proxy_fastcgi for "location / "
use apache as backend for location / , because when you access
http://u.2u4u.com.cn/ , the page contains two links:

1. http://u.2u4u.com.cn/home
2. http://u.2u4u.com.cn/bbs

php-cgi and nginx need an ending slahs so the 'fastcgi_index index.php;' would work.
But i don't know how to write the diretive

when use apache as backend, the ending slash is added by default.

So, u.2u4u.com.cn is a hybrid.

1. In location / , proxy it to apache.
2. for .php, proxy it to fastcgi